requirements.txt
kokoro>=0.9.4
soundfile
sounddevice
numpy
keyboard
pyperclip
pystray
Pillow
